% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA.AIR.df.R
\name{BA.AIR.df}
\alias{BA.AIR.df}
\title{Get the latest update for BA~AIRBUS cointegration trading strategy}
\usage{
BA.AIR.df(fromdate = "2015-01-01")
}
\arguments{
\item{fromdate}{Start date of observation}
}
\value{

}
\description{
Get the latest update for BA~AIRBUS cointegration trading strategy
}
\examples{
BA.AIR.df(fromdate)
}
