% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA.AIR.plot.R
\name{BA.AIR.plot}
\alias{BA.AIR.plot}
\title{Get the latest cointegration plot for BA~AIRBUS trading pair.}
\usage{
BA.AIR.plot(fromdate = "2015-01-01")
}
\arguments{
\item{fromdate}{Start date of observation}
}
\value{

}
\description{
Get the latest cointegration plot for BA~AIRBUS trading pair.
}
\examples{
BA.AIR.plot(fromdate)
}
