% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA.AIR.strat.R
\name{BA.AIR.strat}
\alias{BA.AIR.strat}
\title{Get the latest position suggestion for BA~AIRBUS cointegration trading strategy}
\usage{
BA.AIR.strat(fromdate = "2015-01-01")
}
\arguments{
\item{fromdate}{Start date of observation}
}
\value{

}
\description{
Get the latest position suggestion for BA~AIRBUS cointegration trading strategy
}
\examples{
BA.AIR.strat(fromdate)
}
