% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duocol.api.R
\name{duocol.api}
\alias{duocol.api}
\title{API Query}
\usage{
duocol.api(series_id, APIkey, observation_start)
}
\arguments{
\item{series_id}{Series id of API}

\item{APIkey}{API key}

\item{observation_start}{Observation start date with}
}
\value{

}
\description{
Duo column data
}
\examples{
duocol.api(series_id, APIkey, observation_start)
}
