% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duocol.api.R
\name{duocol.api}
\alias{duocol.api}
\title{API Query}
\usage{
duocol.api(URL, series_id, APIkey, observation_start)
}
\arguments{
\item{URL}{URL}

\item{series_id}{Series id of API}

\item{APIkey}{API key}

\item{observation_start}{Observation start date with}
}
\value{

}
\description{
Duo column data
}
\examples{
duocol.api(URL, series_id, APIkey, observation_start)
duocol.api("http://api.eia.gov/series/","TOTAL.TETCBUS.M","4b7fc3dfb09c8ce47ef1323ea2eb5619","197301")
}
